+++ src/main/java/com/ourmenu/backend/domain/user/dto/request/EmailRequest.java
@@ -1,7 +1,6 @@
-package com.ourmenu.backend.domain.user.dto;
+package com.ourmenu.backend.domain.user.dto.request;
 
 import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -9,13 +8,10 @@ import lombok.NoArgsConstructor;
 
 @Builder
 @Getter
-@AllArgsConstructor
 @NoArgsConstructor
-public class VerifyEmailRequest {
+@AllArgsConstructor
+public class EmailRequest {
 
     @Email
     private String email;
-
-    @NotBlank
-    private String confirmCode;
 }
