+++ src/main/java/com/ourmenu/backend/domain/user/dto/request/VerifyEmailRequest.java
@@ -1,6 +1,7 @@
-package com.ourmenu.backend.domain.user.dto;
+package com.ourmenu.backend.domain.user.dto.request;
 
 import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -8,10 +9,13 @@ import lombok.NoArgsConstructor;
 
 @Builder
 @Getter
-@NoArgsConstructor
 @AllArgsConstructor
-public class EmailRequest {
+@NoArgsConstructor
+public class VerifyEmailRequest {
 
     @Email
     private String email;
+
+    @NotBlank
+    private String confirmCode;
 }
