+++ src/main/java/com/ourmenu/backend/domain/user/dto/response/SignUpResponse.java
@@ -1,0 +1,17 @@
+package com.ourmenu.backend.domain.user.dto.response;
+
+import lombok.Builder;
+import lombok.Data;
+
+
+@Data
+@Builder
+public class SignUpResponse {
+
+    private String grantType;
+    private String accessToken;
+    private String refreshToken;
+    private Long accessTokenExpiredAt;
+
+    private Long refreshTokenExpiredAt;
+}
